Refactor components to include Loader for improved loading states, implemented the layout.js and streamline imports across pages
