+++ frontend/src/pages/AnnouncementsPage.js
@@ -1,10 +1,7 @@
-import React from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
-import { LuLoader2, LuAlertCircle } from 'react-icons/lu';
+import { LuAlertCircle } from 'react-icons/lu';
 import { useAnnouncementsPageData } from '../hooks/useAnnouncementsPageData';
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import AlertCard from '../components/AlertCard';
+import { Loader, AlertCard } from '../components';
 
 const AnnouncementsPage = () => {
   const { data, error, loading } = useAnnouncementsPageData();
@@ -32,39 +29,28 @@ const AnnouncementsPage = () => {
     }
     };
 
-  // Loading state
+  
     if (loading) {
-    return (
-        <div className="min-h-screen bg-gray-50 flex items-center justify-center">
-        <div className="text-center">
-            <LuLoader2 className="w-12 h-12 text-blue-600 animate-spin mx-auto mb-4" />
-            <p className="text-gray-600 text-lg">Loading announcements...</p>
-        </div>
-        </div>
-    );
+      return <Loader isOpen={true} message="Loading announcements..." />
     }
 
   // Error state
   if (error) {
     return (
         <>
-        <Header />
-        <div className="min-h-screen bg-gray-50 flex items-center justify-center">
         <div className="text-center max-w-md mx-auto p-6">
             <LuAlertCircle className="w-16 h-16 text-red-500 mx-auto mb-4" />
             <h2 className="text-2xl font-bold text-gray-800 mb-2">Error Loading Announcements</h2>
             <p className="text-gray-600 mb-4">
             {error.message || 'Failed to load announcements. Please try again later.'}
             </p>
-            <button 
+            <button
             onClick={() => window.location.reload()}
             className="bg-blue-600 text-white px-6 py-2 rounded-lg hover:bg-blue-700 transition-colors"
             >
             Retry
             </button>
         </div>
-        </div>
-        <Footer />
         </>
     );
     }
@@ -74,7 +60,6 @@ const AnnouncementsPage = () => {
 
   return (
    <>
-    <Header />
     <div className="min-h-screen bg-gray-50 py-8 px-4 sm:px-8 lg:px-16">
       <div className="mx-auto">
         {/* Header */}
@@ -121,8 +106,7 @@ const AnnouncementsPage = () => {
         )}
       </div>
     </div>
-    <Footer />
-   </> 
+   </>
   );
 };
