+++ frontend/src/pages/HomePage.js
@@ -1,14 +1,10 @@
 import React from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import HeroSection from "../components/HeroSection";
+import {HeroSection} from "../components";
 
 function HomePage() {
   return (
     <div>
-      <Header />
       <HeroSection />
-      <Footer />
     </div>
   );
 }
