+++ frontend/src/pages/index.js
@@ -1,0 +1,6 @@
+export { default as AllClubsPage } from './AllClubsPage';
+export { default as AnnouncementsPage } from './AnnouncementsPage';
+export { default as ClubPage } from './ClubPage';
+export { default as ContactsPage } from './ContactsPage';
+export { default as EventPage } from './EventPage';
+export { default as HomePage } from './HomePage';
